Add fixes
